add support for multiple payments (whoops)
