+++ src/GPCSV/FieldFactory.php
@@ -7,75 +7,74 @@ use GPCSV\Field;
 class FieldFactory
 {
     // If not specified here, the Standard field type will be used
-    // note that this also dictates the order of fields in CSV body
     /** @var array<string, array<string, array<int, string>|int|string>> */
     public static $fieldDefinitions = [
-        'destinationCountry'          => ['class' => Field\Standard::class, 'minLength' => 2, 'maxLength' => 2],
-        'paymentType'                 => ['class' => Field\ValueList::class, 'values' => ['AT', 'DD', 'LOW', 'MT210', 'URG']],
-        'sendersReference'            => ['class' => Field\Standard::class, 'minLength' => 1, 'maxLength' => 16],
-        'relatedReference'            => ['class' => Field\Standard::class, 'minLength' => 1, 'maxLength' => 16],
-        'valueDate'                   => ['class' => Field\Numeric::class, 'minLength' => 10, 'maxLength' => 10],
-        'debitAccountNumber'          => ['class' => Field\Standard::class],
-        'debitAmount'                 => ['class' => Field\Decimal::class],
-        'creditAmount'                => ['class' => Field\Decimal::class],
-        'currency'                    => ['class' => Field\Standard::class, 'minLength' => 3, 'maxLength' => 3],
-        'draweeCountry'               => ['class' => Field\Standard::class, 'minLength' => 2, 'maxLength' => 2],
-        'originatorName'              => ['class' => Field\Standard::class],
-        'originatorAccount'           => ['class' => Field\Standard::class],
-        'originatorAddress'           => ['class' => Field\Standard::class],
-        'originatorCity'              => ['class' => Field\Standard::class],
-        'originatorCountry'           => ['class' => Field\Standard::class],
-        'beneficiaryAccount'          => ['class' => Field\Standard::class],
-        'beneficiaryName'             => ['class' => Field\Standard::class],
-        'beneficiaryAddress'          => ['class' => Field\Standard::class],
-        'beneficiaryCity'             => ['class' => Field\Standard::class],
-        'beneficiaryPostalCode'       => ['class' => Field\Standard::class],
-        'beneficiaryCountry'          => ['class' => Field\Standard::class, 'minLength' => 2, 'maxLength' => 2],
-        'beneficiaryBankID'           => ['class' => Field\Standard::class],
-        'beneficiaryInformationLine1' => ['class' => Field\Standard::class],
-        'beneficiaryInformationLine2' => ['class' => Field\Standard::class],
-        'beneficiaryInformationLine3' => ['class' => Field\Standard::class],
-        'beneficiaryInformationLine4' => ['class' => Field\Standard::class],
-        'intermediaryBankID'          => ['class' => Field\Standard::class],
-        'receivingBankID'             => ['class' => Field\Unsupported::class],
-        'receiverInformationLine1'    => ['class' => Field\Receiver1::class],
-        'receiverInformationLine2'    => ['class' => Field\Receiver2::class],
-        'receiverInformationLine3'    => ['class' => Field\Receiver2::class],
-        'receiverInformationLine4'    => ['class' => Field\Receiver2::class],
-        'beneficiaryEmail'            => ['class' => Field\Email::class],
-        'notes'                       => ['class' => Field\Standard::class],
-        'sendNoteToBeneficiary'       => ['class' => Field\Numeric::class],
-        'charges'                     => ['class' => Field\ValueList::class, 'values' => ['SHA', 'OUR', 'BEN']],
-        'userDefinedField1'           => ['class' => Field\Standard::class],
-        'userDefinedField2'           => ['class' => Field\Standard::class],
-        'userDefinedField3'           => ['class' => Field\Standard::class],
-        'userDefinedField4'           => ['class' => Field\Standard::class],
-        'userDefinedField5'           => ['class' => Field\Standard::class],
-        'templateCode'                => ['class' => Field\Standard::class],
-        'templateName'                => ['class' => Field\Standard::class],
-        'reserved'                    => ['class' => Field\Unsupported::class],
-        'purposeOfPayment'            => ['class' => Field\Standard::class],
-        'beneficiaryAdvising'         => ['class' => Field\Standard::class],
-        'systemField1'                => ['class' => Field\Standard::class],
-        'systemField2'                => ['class' => Field\Standard::class],
-        'systemField3'                => ['class' => Field\Standard::class],
-        'systemField4'                => ['class' => Field\Standard::class],
-        'systemField5'                => ['class' => Field\Standard::class],
-        'systemField6'                => ['class' => Field\Standard::class],
-        'systemField7'                => ['class' => Field\Standard::class],
-        'systemField8'                => ['class' => Field\Standard::class],
-        'systemField9'                => ['class' => Field\Standard::class],
-        'systemField10'               => ['class' => Field\Standard::class],
-        'systemField11'               => ['class' => Field\Standard::class],
-        'systemField12'               => ['class' => Field\Standard::class],
-        'systemField13'               => ['class' => Field\Standard::class],
-        'systemField14'               => ['class' => Field\Standard::class],
-        'systemField15'               => ['class' => Field\Standard::class],
-        'systemField16'               => ['class' => Field\Standard::class],
-        'systemField17'               => ['class' => Field\Standard::class],
-        'systemField18'               => ['class' => Field\Standard::class],
-        'systemField19'               => ['class' => Field\Standard::class],
-        'systemField20'               => ['class' => Field\Standard::class],
+        'destinationCountry'          => ['class' => Field\Standard::class, 'label' => 'Destination Country', 'minLength' => 2, 'maxLength' => 2],
+        'paymentType'                 => ['class' => Field\ValueList::class, 'label' => 'Payment Type', 'values' => ['AT', 'DD', 'LOW', 'MT210', 'URG']],
+        'sendersReference'            => ['class' => Field\Standard::class, 'label' => 'Sender\'s Reference' ,'minLength' => 1, 'maxLength' => 16],
+        'relatedReference'            => ['class' => Field\Standard::class, 'label' => 'Related Reference' ,'minLength' => 1, 'maxLength' => 16],
+        'valueDate'                   => ['class' => Field\Numeric::class, 'label' => 'Value Date' ,'minLength' => 8, 'maxLength' => 8],
+        'debitAccountNumber'          => ['class' => Field\Standard::class, 'label' => 'Debit Account Number'],
+        'debitAmount'                 => ['class' => Field\Decimal::class, 'label' => 'Debit Amount'],
+        'creditAmount'                => ['class' => Field\Decimal::class, 'label' => 'Credit Amount'],
+        'currency'                    => ['class' => Field\Standard::class, 'label' => 'Currency' ,'minLength' => 3, 'maxLength' => 3],
+        'draweeCountry'               => ['class' => Field\Standard::class, 'label' => 'Drawee Country' ,'minLength' => 2, 'maxLength' => 2],
+        'originatorName'              => ['class' => Field\Standard::class, 'label' => 'Originator Name'],
+        'originatorAccount'           => ['class' => Field\Standard::class, 'label' => 'Originator Account'],
+        'originatorAddress'           => ['class' => Field\Standard::class, 'label' => 'Originator Address'],
+        'originatorCity'              => ['class' => Field\Standard::class, 'label' => 'Originator City'],
+        'originatorCountry'           => ['class' => Field\Standard::class, 'label' => 'Originator Country'],
+        'beneficiaryAccount'          => ['class' => Field\Standard::class, 'label' => 'Beneficiary Account'],
+        'beneficiaryName'             => ['class' => Field\Standard::class, 'label' => 'Beneficiary Name'],
+        'beneficiaryAddress'          => ['class' => Field\Standard::class, 'label' => 'Beneficiary Address'],
+        'beneficiaryCity'             => ['class' => Field\Standard::class, 'label' => 'Beneficiary City'],
+        'beneficiaryPostalCode'       => ['class' => Field\Standard::class, 'label' => 'Beneficiary Postal Code'],
+        'beneficiaryCountry'          => ['class' => Field\Standard::class, 'label' => 'Beneficiary Country' ,'minLength' => 2, 'maxLength' => 2],
+        'beneficiaryBankID'           => ['class' => Field\Standard::class, 'label' => 'Beneficiary Bank ID'],
+        'beneficiaryInformationLine1' => ['class' => Field\Standard::class, 'label' => 'Beneficiary Information Line 1'],
+        'beneficiaryInformationLine2' => ['class' => Field\Standard::class, 'label' => 'Beneficiary Information Line 2'],
+        'beneficiaryInformationLine3' => ['class' => Field\Standard::class, 'label' => 'Beneficiary Information Line 3'],
+        'beneficiaryInformationLine4' => ['class' => Field\Standard::class, 'label' => 'Beneficiary Information Line 4'],
+        'intermediaryBankID'          => ['class' => Field\Standard::class, 'label' => 'Intermediary Bank ID'],
+        'receivingBankID'             => ['class' => Field\Unsupported::class, 'label' => 'Receiving Bank ID'],
+        'receiverInformationLine1'    => ['class' => Field\Receiver1::class, 'label' => 'Receiver Information Line 1'],
+        'receiverInformationLine2'    => ['class' => Field\Receiver2::class, 'label' => 'Receiver Information Line 2'],
+        'receiverInformationLine3'    => ['class' => Field\Receiver2::class, 'label' => 'Receiver Information Line 3'],
+        'receiverInformationLine4'    => ['class' => Field\Receiver2::class, 'label' => 'Receiver Information Line 4'],
+        'beneficiaryEmail'            => ['class' => Field\Email::class, 'label' => 'Beneficiary Email'],
+        'notes'                       => ['class' => Field\Standard::class, 'label' => 'Notes'],
+        'sendNoteToBeneficiary'       => ['class' => Field\Numeric::class, 'label' => 'Send Note to Beneficiary'],
+        'charges'                     => ['class' => Field\ValueList::class, 'label' => 'Charges' ,'values' => ['SHA', 'OUR', 'BEN']],
+        'userDefinedField1'           => ['class' => Field\Standard::class, 'label' => 'User Defined Field 1'],
+        'userDefinedField2'           => ['class' => Field\Standard::class, 'label' => 'User Defined Field 2'],
+        'userDefinedField3'           => ['class' => Field\Standard::class, 'label' => 'User Defined Field 3'],
+        'userDefinedField4'           => ['class' => Field\Standard::class, 'label' => 'User Defined Field 4'],
+        'userDefinedField5'           => ['class' => Field\Standard::class, 'label' => 'User Defined Field 5'],
+        'templateCode'                => ['class' => Field\Standard::class, 'label' => 'Template Code'],
+        'templateName'                => ['class' => Field\Standard::class, 'label' => 'Template Name'],
+        'reserved'                    => ['class' => Field\Unsupported::class, 'label' => 'Reserved'],
+        'purposeOfPayment'            => ['class' => Field\Standard::class, 'label' => 'Purpose of Payment'],
+        'beneficiaryAdvising'         => ['class' => Field\Standard::class, 'label' => 'Beneficiary Advising'],
+        'systemField1'                => ['class' => Field\Standard::class, 'label' => 'System Field 1'],
+        'systemField2'                => ['class' => Field\Standard::class, 'label' => 'System Field 2'],
+        'systemField3'                => ['class' => Field\Standard::class, 'label' => 'System Field 3'],
+        'systemField4'                => ['class' => Field\Standard::class, 'label' => 'System Field 4'],
+        'systemField5'                => ['class' => Field\Standard::class, 'label' => 'System Field 5'],
+        'systemField6'                => ['class' => Field\Standard::class, 'label' => 'System Field 6'],
+        'systemField7'                => ['class' => Field\Standard::class, 'label' => 'System Field 7'],
+        'systemField8'                => ['class' => Field\Standard::class, 'label' => 'System Field 8'],
+        'systemField9'                => ['class' => Field\Standard::class, 'label' => 'System Field 9'],
+        'systemField10'               => ['class' => Field\Standard::class, 'label' => 'System Field 10'],
+        'systemField11'               => ['class' => Field\Standard::class, 'label' => 'System Field 11'],
+        'systemField12'               => ['class' => Field\Standard::class, 'label' => 'System Field 12'],
+        'systemField13'               => ['class' => Field\Standard::class, 'label' => 'System Field 13'],
+        'systemField14'               => ['class' => Field\Standard::class, 'label' => 'System Field 14'],
+        'systemField15'               => ['class' => Field\Standard::class, 'label' => 'System Field 15'],
+        'systemField16'               => ['class' => Field\Standard::class, 'label' => 'System Field 16'],
+        'systemField17'               => ['class' => Field\Standard::class, 'label' => 'System Field 17'],
+        'systemField18'               => ['class' => Field\Standard::class, 'label' => 'System Field 18'],
+        'systemField19'               => ['class' => Field\Standard::class, 'label' => 'System Field 19'],
+        'systemField20'               => ['class' => Field\Standard::class, 'label' => 'System Field 20'],
     ];
 
     public static function createField(string $name): Field
@@ -86,8 +85,9 @@ class FieldFactory
 
         $definition = self::$fieldDefinitions[$name];
         $class      = $definition['class'];
+        assert(is_string($class));
 
-        $field = new $class(array_diff_key($definition, ['class' => $class]));
+        $field = new $class(['name' => $name] + array_diff_key($definition, ['class' => $class]));
 
         if (!($field instanceof Field)) {
             throw new Exception('Class could not be instantiated: ' . $class);
